DTSAM-1210 Move the delete booking controller into testingsupport.
