+++ src/test/java/uk/gov/hmcts/reform/judicialbooking/controller/endpoints/DeleteBookingControllerTest.java
@@ -6,6 +6,7 @@ import org.mockito.InjectMocks;
 import org.mockito.MockitoAnnotations;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import uk.gov.hmcts.reform.judicialbooking.controller.endpoints.testingsupport.DeleteBookingController;
 import uk.gov.hmcts.reform.judicialbooking.domain.service.BookingOrchestrator;
 
 import java.util.UUID;
